pyproject.toml
[tool.poetry]
name = "cinna-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MPL-2.0"
readme = "README.md"
packages = [{include = "cinna_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.8.1"
python-dotenv = "^1.0.0"
tinydb = "^4.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
